fix: check if directory exists

+++ backuprotate.sh
@@ -128,7 +128,9 @@ function prune () {
    fi
    rm -f /tmp/.backuprotate.tmp
  else
-   find "$base/$dir" -mindepth 1 -maxdepth 1 -type d | sort -rn | awk " NR > $keep" | while read f; do rm -rf "$f"; done
+   if [ -d "$base/$dir" ]; then
+     find "$base/$dir" -mindepth 1 -maxdepth 1 -type d | sort -rn | awk " NR > $keep" | while read f; do rm -rf "$f"; done
+   fi
  fi
 }
